API Reference

WiNDCHousehold.HouseholdTable

WiNDCHousehold.NYSE_data

WiNDCHousehold.absorption

WiNDCHousehold.adjust_capital_demand

WiNDCHousehold.adjusted_capital_income

WiNDCHousehold.adjusted_consumption

WiNDCHousehold.adjusted_wages

WiNDCHousehold.aggregate_transfer_payment

WiNDCHousehold.armington_data

WiNDCHousehold.average_labor_tax_rate

WiNDCHousehold.build_household_table

WiNDCHousehold.build_transfer_payments

WiNDCHousehold.calibration_model_1

WiNDCHousehold.calibration_model_2

WiNDCHousehold.capital_stock_data

WiNDCHousehold.capital_tax_rate

WiNDCHousehold.clean_aggregate_medi_data

WiNDCHousehold.clean_cps_data_year

WiNDCHousehold.compute_economy_wide_savings

WiNDCHousehold.compute_savings_markup_consumption

WiNDCHousehold.compute_savings_markup_save

WiNDCHousehold.compute_savings_rate

WiNDCHousehold.consumption_data

WiNDCHousehold.cpi_data

WiNDCHousehold.cps_vs_nipa_income_categories

WiNDCHousehold.create_personal_consumption

WiNDCHousehold.disposition_data

WiNDCHousehold.duty_rate

WiNDCHousehold.fetch_zip_data

WiNDCHousehold.fica_tax_rate

WiNDCHousehold.get_capital_ownership

WiNDCHousehold.get_cps_data_api

WiNDCHousehold.government_data

WiNDCHousehold.government_deficit

WiNDCHousehold.initialize_table

WiNDCHousehold.invest_data

WiNDCHousehold.labor_endowment_multiplier

WiNDCHousehold.labor_supply

WiNDCHousehold.leisure_data

WiNDCHousehold.leisure_demand

WiNDCHousehold.load_acs_data_api

WiNDCHousehold.load_acs_medicare_data

WiNDCHousehold.load_capital_tax_rates

WiNDCHousehold.load_cps_data_api

WiNDCHousehold.load_labor_tax_rates

WiNDCHousehold.load_medicare_data

WiNDCHousehold.load_nipa_data_api

WiNDCHousehold.load_state_fips

WiNDCHousehold.margin_supply_demand

WiNDCHousehold.marginal_labor_tax_rate

WiNDCHousehold.netports

WiNDCHousehold.nipa_fringe_benefit_markup

WiNDCHousehold.other_income

WiNDCHousehold.output_tax_rate

WiNDCHousehold.regional_local_supply

WiNDCHousehold.regional_national_supply

WiNDCHousehold.representative_agent_data

WiNDCHousehold.saverate_data

WiNDCHousehold.sectoral_output

WiNDCHousehold.ssk_data

WiNDCHousehold.tax_rate

WiNDCHousehold.total_supply

WiNDCHousehold.trans_data

WiNDCHousehold.transfer_weights

WiNDCHousehold.windc_vs_nipa_income_categories

WiNDCHousehold.HouseholdTable — Type

Household

The primary container for WiNDC household disaggregation data tables. There are three fields, all dataframes:

data: The main data table.

sets: The sets table, describing the different sets used in the model.

elements: The elements table, describing the different elements in the model.

WiNDCHousehold.NYSE_data — Method

NYSE_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts NYSE-related parameters from the regional data table.

    table(data, 
        :Household_Supply,
        :Capital_Demand;
        normalize=:Use
    )

WiNDCHousehold.absorption — Method

absorption(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value
)

Calculate the absorption for each commodity in each region. The absorption is defined as the sum of:

Intermediate_Demand

Other_Final_Demand

Note that absorption is negative.

WiNDCHousehold.adjust_capital_demand — Method

adjust_capital_demand(
    HH::HouseholdTable,
    state_table::State,
    raw_data::RawHouseholdData;
)

Convert gross capital demand in the state_table to net capital demand by removing the capital tax based on state-level capital tax rates provided in raw_data.

\[ {\rm Net\ Capital\ Demand} = \frac{\rm Gross\ Capital\ Demand}{1 + r}\]

Raw Data

raw_data.capital_tax_rates (State-level capital tax rates)

Parameters Disaggregated

Capital_Demand

Sets Added

capital_tax

Domain: row

Description: "Capital Tax"

Elements: cap_tax

Capital_Tax - From state_table

WiNDCHousehold.adjusted_capital_income — Method

adjusted_capital_income(
    HH::HouseholdTable,
    state_table::WiNDCRegional.State,
    raw_data::RawHouseholdData
)

Compute the adjusted capital income for each household in HH based on capital ownership and total capital demand.

Arguments

HH::HouseholdTable: The household table.

state_table::WiNDCRegional.State: The state-level WiNDC table.

raw_data::RawHouseholdData: The raw household data containing CPS data.

Returns

adjusted_capital::DataFrame: A DataFrame with columns :hh, :year, :state, and :capital representing the adjusted capital income for each household.

Computation

The adjusted capital income is computed using the formula:

\[ {\rm Adjusted\ Capital\ Income} = {\rm Capital\ Ownership\ Share} \cdot {\rm Total\ Capital\ Demand} \cdot \frac{{\rm Interest\ Income}}{\sum_{hh}{\rm Interest\ Income}}\]

WiNDCHousehold.adjusted_consumption — Method

adjusted_consumption(
    HH::HouseholdTable,
    state_table::WiNDCRegional.State,
    raw_data::RawHouseholdData
)

Compute the adjusted consumption for each household in HH based on consumption shares, personal consumption from state_table, and the household savings rate.

Arguments

HH::HouseholdTable: The household table.

state_table::WiNDCRegional.State: The state-level WiNDC table.

raw_data::RawHouseholdData: The raw household data containing CPS data.

Returns

adjusted_consumption::DataFrame: A DataFrame with columns :hh, :year, :state, and :consumption representing the adjusted consumption for each household.

Computation

The adjusted consumption is computed using the formula:

\[ {\rm Adjusted\ Consumption} = \frac{{\rm Consumption\ Share} \cdot {\rm Personal\ Consumption}}{1 - {\rm Savings\ Rate}}\]

WiNDCHousehold.adjusted_wages — Method

adjusted_wages(
    HH::HouseholdTable,
    state_table::WiNDCRegional.State,
    raw_data::RawHouseholdData;
    tax_adjustment::Bool = true
)

Compute the adjusted wages for each household in HH based on labor endowment multipliers and labor tax rates.

Arguments

HH::HouseholdTable: The household table.

state_table::WiNDCRegional.State: The state-level WiNDC table.

raw_data::RawHouseholdData: The raw household data containing CPS data and labor tax rates.

Keyword Arguments

tax_adjustment::Bool = true: Whether to apply labor tax adjustments.

Returns

adjusted_wages::DataFrame: A DataFrame with columns :hh, :year, :state, and :wage representing the adjusted wages for each household.

Computation

The adjusted wages are computed using the formula:

\[ {\rm Adjusted\ Wage} = {\rm Wage\ Share} \cdot ({\rm tax\ adjustment}\cdot {\rm Labor\ Tax\ Rate} - 1) \cdot {\rm Labor\ Endowment\ Multiplier} \cdot \sum_{sectors}{\rm Labor\ Demand}\]

WiNDCHousehold.aggregate_transfer_payment — Method

aggregate_transfer_payment(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value,
    parameter = :total_transfers,
)

Aggregate the transfer payments for each household type in each region.

Required Arguments

HH::HouseholdTable: The household data.

Keyword Arguments

column::Symbol = :value: The column to be used for the aggregation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:total_transfers`: The name of the parameter column.

Returns

A DataFrame with the aggregated transfer payments for each household type in each region.

Calculation

The total transfer payments are calculated as:

\[{\rm Total\_Transfers} = \sum_{\rm transfers} {\rm Transfer\_Payment}\]

WiNDCHousehold.armington_data — Method

armington_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts Armington-related parameters from the regional data table.

    vcat(
        absorption(data; normalize = true),
        table(data, :Reexport, :National_Demand, :Local_Demand, :Import, :Margin_Demand; normalize = :Use),
        tax_rate(data),
        duty_rate(data)
    )

Aggregate data

WiNDCHousehold.absorption

WiNDCHousehold.tax_rate

WiNDCHousehold.duty_rate

WiNDCHousehold.average_labor_tax_rate — Method

average_labor_tax_rate(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value,
    parameter = :average_labor_tax_rate,
)

Calculate the average labor tax rate for each household type in each region. The average labor tax rate is defined as the ratio of Average_Labor_Tax to Labor_Endowment.

Required Arguments

HH::HouseholdTable: The household data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:averagelabortax_rate`: The name of the parameter column.

minimal::Bool = false: Whether to return a minimal output. If true, only the essential columns are returned: [:col, :region, :year, :parameter, output].

Returns

A DataFrame with the average labor tax rate for each household type in each region.

Calculation

The average labor tax rate is calculated as:

\[{\rm Average\_Labor\_Tax\_Rate} = \frac{\rm Average\_Labor\_Tax}{\sum_{\rm Destinations} \rm Labor\_Endowment}\]

WiNDCHousehold.build_household_table — Method

build_household_table(
        state_table::State,
        raw_data::RawHouseholdData;
    )

Build a HouseholdTable from the given state_table and raw_data.

Disaggregation Steps

The disaggregation is performed through the following steps:

WiNDCHousehold.initialize_table

WiNDCHousehold.adjust_capital_demand

WiNDCHousehold.build_transfer_payments

WiNDCHousehold.build_transfer_payments — Method

build_transfer_payments(
    HH::HouseholdTable,
    raw_data::RawHouseholdData,
)

Add transfer payments to the HouseholdTable HH based on the data in raw_data.

Raw Data

WiNDCHousehold.transfer_weights

raw_data.income (CPS income data)

raw_data.medicare (Medicare and Medicaid data)

raw_data.income_categories (CPS income categories)

Sets Added

Transfer_Payment

Domain: parameter

Description: "Transfer Payments"

Elements: transfer_payment

transfer_payments

Domain: row

Description: "Transfer Payments"

Element	Description

hucval	unemployment compensation

hwcval	workers compensation

hssval	social security

hssival	supplemental security

hpawval	public assistance or welfare

hvetval	veterans benefits

hsurval	survivors income

hdisval	disability

hedval	educational assistance

hcspval	child support

hfinval	financial assistance

medicare	edicare

medicaid	edicaid

other	Other Income

The year for medicare and medicaid is fixed at 2024. Need to update for more years in the future.

WiNDCHousehold.calibration_model_1 — Method

calibration_model_1(HH::HouseholdTable, state_table::State, HH_Raw_Data::RawHouseholdData)

This is a placeholder docstring. It has been intentionally left blank as this model can be simplified further.

To Do:

Add errors if model fails to solve

WiNDCHousehold.calibration_model_2 — Method

calibration_model_2(
    HH::HouseholdTable, 
    state_table::State, 
    HH_Raw_Data::RawHouseholdData,
    calibration_model_1::JuMP.Model
)

This is a placeholder docstring. It has been intentionally left blank as this model can be simplified further.

To Do:

Add errors if model fails to solve

WiNDCHousehold.capital_stock_data — Method

capital_stock_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts capital stock-related parameters from the regional data table.

    table(data, :Capital_Demand; normalize = :Use)

WiNDCHousehold.capital_tax_rate — Method

capital_tax_rate(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value,
    parameter = :capital_tax_rate,
)

Calculate the capital tax rate for each household type in each region. The capital tax rate is defined as the ratio of Capital_Tax to Capital_Demand.

Required Arguments

HH::HouseholdTable: The household data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:capitaltaxrate`: The name of the parameter column.

Returns

A DataFrame with the capital tax rate for each household type in each region.

Calculation

The capital tax rate is calculated as:

\[{\rm Capital\_Tax\_Rate} = \frac{\rm Capital\_Tax}{\rm Capital\_Demand}\]

WiNDCHousehold.clean_aggregate_medi_data — Method

clean_aggregate_medi_data(df::DataFrame)

Cleans Medicare or Medicaid aggregate data DataFrame.

WiNDCHousehold.clean_cps_data_year — Method

clean_cps_data_year(year, cps_rw, variables, api_key; state_fips = load_state_fips())

Clean and process the CPS data for a given year.

Arguments

year::Int: The year of the CPS data to load.

cps_rw::Vector{String}: The list of CPS raw variables to load

variables::Vector{String}: The list of CPS variables to process.

api_key::String: The Census API key.

Optional Arguments

state_fips::DataFrame: A DataFrame containing state FIPS codes. Defaults to loading from load_state_fips().

Return

Returns a NamedTuple with the following DataFrames:

income::DataFrame: Income data by household type and state.

count::DataFrame: Count of households by household type and state.

numhh::DataFrame: Number of households by household type and state.

WiNDCHousehold.compute_economy_wide_savings — Method

compute_economy_wide_savings(
        state_table::WiNDCRegional.State
    )

Compute the economy-wide savings rate based on personal consumption and investment from the state_table.

Returns

economy_wide_savings::DataFrame: A DataFrame with columns :year and :economy_wide_sr0 representing the economy-wide savings rate for each year.

Computation

\[ {\rm Economy\ Wide\ Savings\ Rate} = \frac{\rm Investment\ Final\ Demand}{\rm Personal\ Consumption + Investment\ Final\ Demand}\]

WiNDCHousehold.compute_savings_markup_consumption — Method

compute_savings_markup_consumption(
    state_table::WiNDCRegional.State,
    raw_data::RawHouseholdData
)

Compute the savings markup based on personal consumption from the state_table and CPS data in raw_data.

Arguments

state_table::WiNDCRegional.State: The state-level WiNDC table.

raw_data::RawHouseholdData: The raw household data containing CPS data.

Returns

savings_markup_consumption::Float64: The computed savings markup based on consumption.

Computation

\[ {\rm Savings\ Markup} = \frac{\rm Total\ Personal\ Consumption}{\rm Total\ Consumption}\]

WiNDCHousehold.compute_savings_markup_save — Method

savings_markup_save(
    HH::HouseholdTable,
    raw_data::RawHouseholdData
)

Compute the savings markup for the household table HH based on the raw CPS data in raw_data.

Arguments

HH::HouseholdTable: The household table.

raw_data::RawHouseholdData: The raw household data containing CPS data.

Returns

savings_markup_save::Float64: The computed savings markup.

Raw Data Used

raw_data.cps_data (CPS interest data)

WiNDCHousehold.get_capital_ownership

Computation

\[ {\rm Savings\ Markup} = \frac{\rm Gross\ Capital\ Demand}{\rm Total\ Interest}\]

WiNDCHousehold.compute_savings_rate — Method

compute_savings_rate(
        HH::HouseholdTable,
        state_table::WiNDCRegional.State,
        raw_data::RawHouseholdData
    )

Compute the savings rate for each household in HH based on the economy-wide savings rate and the household's interest income and consumption.

Arguments

HH::HouseholdTable: The household table.

state_table::WiNDCRegional.State: The state-level WiNDC table.

raw_data::RawHouseholdData: The raw household data containing CPS data.

Returns

savings_rate::DataFrame: A DataFrame with columns :hh, :year, :state, and :save_rate representing the savings rate for each household.

Computation

This computation is broken into several steps:

Compute the economy-wide savings rate using WiNDCHousehold.compute_economy_wide_savings.

Compute the savings markup for savings and consumption using:

WiNDCHousehold.compute_savings_markup_save and

WiNDCHousehold.compute_savings_markup_consumption.

Calculate the interest shares and total shares for each household.

Finally, compute the savings rate for each household using the formula:

\[ {\rm Savings\ Rate}_{hh} = \frac{\rm (Economy\ Wide\ Savings\ Rate) \cdot (Interest\ Share)}{Total\ Share}}\]

WiNDCHousehold.consumption_data — Method

consumption_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts consumption-related parameters from the regional data table.

    table(data, :Personal_Consumption; normalize = :Use)

WiNDCHousehold.cpi_data — Method

cpi_data(HH::HouseholdTable; output = :DataFrame)

Extracts consumer price index-related parameters from the household data table.

    table(HH, :Personal_Consumption; normalize=:Use) |>
        x -> groupby(x, [:col, :region, :parameter]) |>
        x -> combine(x, :value => sum => :value)

WiNDCHousehold.cps_vs_nipa_income_categories — Method

cps_vs_nipa_income_categories(cps_income::DataFrame, nipa::DataFrame)

Compare CPS income data to NIPA data by income categories.

WiNDCHousehold.create_personal_consumption — Method

create_personal_consumption(
        HH::HouseholdTable,
        state_table::WiNDCRegional.State,
        raw_data::RawHouseholdData,
        M2::JuMP.Model
    )

WiNDCHousehold.disposition_data — Method

disposition_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts disposition-related parameters from the regional data table.

    vcat(
        regional_local_supply(data),
        netports(data),
        total_supply(data),
        regional_national_supply(data)
        )

Aggregate Data

WiNDCHousehold.regional_local_supply

WiNDCHousehold.netports

WiNDCHousehold.total_supply

WiNDCHousehold.regional_national_supply

WiNDCHousehold.duty_rate — Method

dutyrate( HH::HouseholdTable; column::Symbol = :value, output::Symbol = :value, parameter = :dutyrate, minimal::Bool = false )

Calculate the duty rate for each commodity in each region. The duty rate is defined as the ratio of Duty to Import.

Required Arguments

HH::HouseholdTable: The regional data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:duty_rate`: The name of the parameter column.

minimal::Bool = false: Whether to return a minimal output. If true, only the essential columns are returned: [:col, :region, :year, :parameter, output].

WiNDCHousehold.fetch_zip_data — Method

fetch_zip_data(
    url::String,
    filter_function::Function;;
    output_path::String = tempname(),
)

Download a zip file from a given url and extract the files in the zip file that are in the data NamedTuple.

This function will throw an error if not all files in data are extracted.

Required Arguments

url::String: The url of the zip file to download.

filter_function::Function;: A function that takes a string and returns a boolean. This function is used to filter the files in the zip file, it should return true if the file should be extracted and false otherwise.

Optional Arguments

output_path::String: The path to save the extracted files. Default is a

temporary directory. If this is not an absolute path, it will be joined with the current working directory.

Output

Returns a vector of the absolute paths to the extracted files.

WiNDCHousehold.fica_tax_rate — Method

fica_tax_rate(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value,
    parameter = :fica_tax_rate,
)

Calculate the FICA tax rate for each household type in each region. The FICA tax rate is defined as the ratio of FICA_Tax to Labor_Endowment.

Required Arguments

HH::HouseholdTable: The household data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:ficataxrate`: The name of the parameter column.

minimal::Bool = false: Whether to return a minimal output. If true, only the essential columns are returned: [:col, :region, :year, :parameter, output].

Returns

A DataFrame with the FICA tax rate for each household type in each region.

Calculation

The FICA tax rate is calculated as:

\[{\rm FICA\_Tax\_Rate} = \frac{\rm FICA\_Tax}{\sum_{\rm Destinations} \rm Labor\_Endowment}\]

WiNDCHousehold.get_capital_ownership — Method

get_capital_ownership(
    raw_data::RawHouseholdData;
    type::Symbol = :all
)

Get the capital ownership share from raw_data.

Arguments

raw_data::RawHouseholdData: The raw household data containing CPS and NIPA data.

Keyword Arguments

type::Symbol = :all: The type of capital ownership to retrieve. Options are:

:all: Returns 1 (100% ownership).

:partial: Computes the ownership share based on CPS vs NIPA data.

Returns

cap_own::Float64: The capital ownership share.

WiNDCHousehold.get_cps_data_api — Method

get_cps_data_api(year, vars, api_key)

Pull the raw data directly from the API

WiNDCHousehold.government_data — Method

government_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts government-related parameters from the regional data table.

    vcat(
        table(data, 
            :Government_Final_Demand,
            :Transfer_Payment,
            :Labor_Endowment;
            normalize=:Use
        ),
        WiNDCHousehold.government_deficit(data),
        WiNDCHousehold.average_labor_tax_rate(data),
    )

WiNDCHousehold.government_deficit — Method

government_deficit(
    HH::HouseholdTable;
    column::Symbol = :value,
    output::Symbol = :value,
    parameter::Symbol = :government_deficit,
)

Calculate the government deficit for each year.

Arguments

HH::HouseholdTable: The household table containing household data.

Keyword Arguments

column::Symbol: The column to aggregate (default is :value).

output::Symbol: The name of the output column (default is :value).

parameter::Symbol: The parameter to assign to the output (default is :government_deficit).

Returns

A DataFrame with the government deficit for each year.

Calculation

The government deficit is calculated as:

\[{\rm Government\_Deficit} = - \sum (Government\_Final\_Demand - Transfer\_Payment - Average\_Labor\_Tax - FICA\_Tax - Capital\_Tax - Output\_Tax - Tax - Duty)\]

WiNDCHousehold.initialize_table — Method

initialize_table(state_table::State)

Create an initial HouseholdTable from a State table by selecting only the relevant parameters, sets, and elements.

Essentially, all parameters, sets, and elements are maintained except those related to personal_consumption.

Parameters Maintained

Capital_Demand

Duty

Export

Government_Final_Demand

Household_Supply

Import

Intermediate_Demand

Intermediate_Supply

Investment_Final_Demand

Labor_Demand

Local_Demand

Local_Margin_Supply

Margin_Demand

National_Demand

National_Margin_Supply

Output_Tax

Reexport

Tax

Sets Maintained

duty

export

government_final_demand

import

investment_final_demand

local_demand

margin

national_demand

reexport

sector

tax

trade

transport

state

capital_demand

commodity

labor_demand

output_tax

year

Final_Demand

Other_Final_Demand

Margin_Supply

Use

Supply

Value_Added

Sets Added

household

Domain: col

Description: "Household Categories"

Elements: hh1, hh2, hh3, hh4, hh5

WiNDCHousehold.invest_data — Method

invest_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts investment-related parameters from the regional data table.

    table(data, 
        :Investment_Final_Demand,
        :Savings;
        normalize=:Use
    )

WiNDCHousehold.labor_endowment_multiplier — Method

labor_endowment_multiplier(
    HH::HouseholdTable,
    state_table::WiNDCRegional.State,
    raw_data::RawHouseholdData
)

Compute the labor endowment multiplier for each region based on labor demand and wages.

Arguments

HH::HouseholdTable: The household table.

state_table::WiNDCRegional.State: The state-level WiNDC table.

raw_data::RawHouseholdData: The raw household data containing CPS data.

Returns

labor_endow_multiplier::DataFrame: A DataFrame with columns :region and :le_mult representing the labor endowment multiplier for each region.

Computation

The labor endowment multiplier is computed using the formula:

\[ {\rm Labor\ Endowment\ Multiplier} = \frac{\rm Fringe\ Benefit\ Markup \cdot \rm Wage}{\rm Labor\ Demand}\]

Changes also made based on commuting patterns to ensure that regions without commuting workers do not have their labor endowment multipliers adjusted beyond 1.

WiNDCHousehold.labor_supply — Method

labor_supply(
    HH::HouseholdTable;
    column::Symbol = :value,
    output::Symbol = :value,
    parameter::Symbol = :labor_supply,
)

Calculate the labor supply for households. The labor supply is defined as the labor endowment adjusted for marginal labor taxes and FICA taxes.

Arguments

HH::HouseholdTable: The household table containing household data.

Keyword Arguments

column::Symbol: The column to aggregate (default is :value).

output::Symbol: The name of the output column (default is :value).

parameter::Symbol: The parameter to assign to the output (default is :labor_supply).

Returns

A DataFrame with the aggregated labor supply for each region and year.

Calculation

Labor supply is calculated as:

\[{\rm Labor\_Supply} = \sum (Labor\_Endowment - Marginal\_Labor\_Tax - FICA\_Tax)\]

WiNDCHousehold.leisure_data — Method

leisure_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts leisure-related parameters from the regional data table.

   vcat(
        table(data, :Labor_Endowment),
        WiNDCHousehold.labor_supply(data),
        WiNDCHousehold.marginal_labor_tax_rate(data),
        WiNDCHousehold.fica_tax_rate(data)
    )

WiNDCHousehold.leisure_demand — Method

leisure_demand(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value,
    parameter = :leisure_demand,
)

Calculate the leisure demand for households. The leisure demand is defined as a fraction of the leisure supply, based on a labor supply income elasticity, which is set to 0.05.

Arguments

HH::HouseholdTable: The household table containing household data.

Keyword Arguments

column::Symbol: The column to aggregate (default is :value).

output::Symbol: The name of the output column (default is :value).

parameter::Symbol: The parameter to assign to the output (default is :leisure_demand).

Returns

A DataFrame with the calculated leisure demand for each region and year.

Calculation

Leisure demand is calculated as:

\[{\rm Leisure\_Demand} = \epsilon \cdot {\rm Leisure\_Supply}\]

where ( \epsilon ) is the labor supply income elasticity (0.05).

WiNDCHousehold.load_acs_data_api — Method

load_acs_data_api(year::Int, census_api_key; output_path::String = tempname())

Pull the raw ACS commuting flow data directly from the Census website for a given year.

Return a DataFrame with the commuting flow data for the specified year, enriched with income data from the CPS.

Arguments

year::Int: The year of the ACS data to load.

census_api_key::String: The Census API key.

Optional Arguments

output_path::String: The path to save the downloaded ACS data. Default is a temporary directory.

WiNDCHousehold.load_acs_medicare_data — Method

load_acs_medicare_data(
    census_api_key::String;
    years::Vector{Int} = 2009:2024,
    state_fips::DataFrame = load_state_fips()
)

Load ACS Medicare data from the Census API for the specified years.

Required Arguments

census_api_key::String: The Census API key.

Optional Arguments

years::Vector{Int}: The years of the ACS data to load. Default is 2009 to 2024.

state_fips::DataFrame: A DataFrame containing state FIPS codes. Default is the result of load_state_fips().

Excluded 2020

The year 2020 is excluded from the default years because the ACS data for that year does not include Medicare data due to low response rates.

Early years?

Currently we don't have data before 2009, so requesting years before. We should update this to just use 2009 data for earlier years.

Return

Returns a DataFrame with shares of income groups for Medicare enrollment by state and year.

WiNDCHousehold.load_capital_tax_rates — Method

load_capital_tax_rates(
    path::String;
    state_abbreviations = WiNDCHousehold.load_state_fips(cols_to_keep = [:state, :abbreviation])
)

Load capital tax rates from a CSV file located at path. The CSV is expected to contain state-level capital tax rate information, which will be matched with state abbreviations provided by state_abbreviations.

WiNDCHousehold.load_cps_data_api — Method

load_cps_data_api(api_key; years = 2001:2022)

Load all the CPS data in the given year range.

WiNDCHousehold.load_labor_tax_rates — Method

load_labor_tax_rates(
        path::String;
        state_abbreviations = WiNDCHousehold.load_state_fips(cols_to_keep = [:state, :abbreviation])
    )

Load labor tax rates from a CSV file located at path. The CSV is expected to contain state-level labor tax rate information, which will be matched with state abbreviations provided by state_abbreviations.

WiNDCHousehold.load_medicare_data — Method

load_medicare_data(path::String; years::UnitRange{Int}=2009:2024)

Load Medicare or Medicaid data from a CSV file at the given path.

Required Arguments

path::String: The path to the CSV file.

Optional Arguments

years::UnitRange{Int}: The range of years to include in the data. Default is 2009 to 2024.

WiNDCHousehold.load_nipa_data_api — Method

load_nipa_data_api(years::Vector{Int}, bea_api_key)

Pull the raw data directly from the API

Return a DataFrame with the NIPA data for the specified years.

Arguments

years::Vector{Int}: The years of the NIPA data to load.

bea_api_key::String: The BEA API key.

WiNDCHousehold.load_state_fips — Method

load_state_fips(;
        path = joinpath(@__DIR__, "maps", "state_fips.csv"),
        cols_to_keep = [:fips, :state]
    )

Load a CSV file containing state FIPS codes and state names. The default path is set to maps/state_fips.csv relative to this file's directory. You can specify which columns to keep using the cols_to_keep argument.

Returns a DataFrame with the specified columns as Strings.

Optional Arguments

path::String: Path to the state FIPS CSV file.

cols_to_keep::Vector{Symbol}: Columns to keep from the CSV file. Default is [:fips, :state].

It is recommended to keep the columns :fips and :state for proper mapping.

WiNDCHousehold.margin_supply_demand — Method

margin_supply_demand(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts margin supply and demand parameters from the regional data table.

    table(data, :Margin_Demand, :Margin_Supply; normalize = :Use)

WiNDCHousehold.marginal_labor_tax_rate — Method

marginal_labor_tax_rate(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value,
    parameter = :marginal_labor_tax_rate,
    minimal::Bool = false
)

Calculate the marginal labor tax rate for each household type in each region. The marginal labor tax rate is defined as the ratio of Marginal_Labor_Tax to Labor_Endowment.

Required Arguments

HH::HouseholdTable: The household data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:marginallabortax_rate`: The name of the parameter column.

minimal::Bool = false: Whether to return a minimal output. If true, only the essential columns are returned: [:col, :region, :year, :parameter, output].

Returns

A DataFrame with the marginal labor tax rate for each household type in each region.

Calculation

The marginal labor tax rate is calculated as:

\[{\rm Marginal\_Labor\_Tax\_Rate} = \frac{\rm Marginal\_Labor\_Tax}{\sum_{\rm Destinations} \rm Labor\_Endowment}\]

WiNDCHousehold.netports — Method

netports(
    HH::HouseholdTable;
    column::Symbol = :value,
    output::Symbol = :value,
    parameter::Symbol = :netport,
    minimal::Bool = false
)

Calculate the netports for each commodity in each region. The netport is defined as the difference between Export and Reexport.

WiNDCHousehold.nipa_fringe_benefit_markup — Method

nipa_fringe_benefit_markup(nipa::DataFrame)

Calculate the fringe benefit markup from NIPA data.

WiNDCHousehold.other_income — Method

other_income(
    HH::HouseholdTable,
    state_table::WiNDCRegional.State,
    raw_data::RawHouseholdData
)

Compute the other income for each household in HH based on adjusted consumption, adjusted wages, transfer payments, and adjusted capital income.

Arguments

HH::HouseholdTable: The household table.

state_table::WiNDCRegional.State: The state-level WiNDC table.

raw_data::RawHouseholdData: The raw household data containing CPS data.

Returns

other_income::DataFrame: A DataFrame with columns :hh, :year, :state, and :income representing the other income for each household.

Computation

The other income is computed using the formula:

\[ {\rm Other\ Income} = {\rm Adjusted\ Consumption} + {\rm Adjusted\ Wages} - {\rm Average\ Transfer\ Payments} - {\rm Adjusted\ Capital\ Income}\]

where:

Adjusted Consumption is obtained from WiNDCHousehold.adjusted_consumption.

Adjusted Wages is obtained from WiNDCHousehold.adjusted_wages.

Average Transfer Payments the average over households and states of the Transfer_Payment parameter in HH.

Adjusted Capital Income is obtained from WiNDCHousehold.adjusted_capital_income.

WiNDCHousehold.output_tax_rate — Method

output_tax_rate(
        HH::HouseholdTable; 
        column::Symbol = :value, 
        output::Symbol = :value,
        parameter = :output_tax_rate,
        minimal::Bool = false
    )

Calculate the output tax rate for each commodity in each region. The output tax rate is defined as the ratio of Output_Tax to the sum over commodities of Intermediate_Supply.

Required Arguments

HH::HouseholdTable: The regional data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:outputtaxrate`: The name of the parameter column.

minimal::Bool = false: Whether to return a minimal output. If true, only the essential columns are returned: [:col, :region, :year, :parameter, output].

WiNDCHousehold.regional_local_supply — Method

regional_local_supply(
    HH::HouseholdTable;
    column::Symbol = :value,
    output::Symbol = :value,
    parameter::Symbol = :region_local_supply,
    minimal::Bool = false
)

Calculate the regional local supply for each commodity in each region. The regional local supply is defined as the sum of Local_Margin_Supply and Local_Demand.

WiNDCHousehold.regional_national_supply — Method

regional_national_supply(
    HH::HouseholdTable;
    column::Symbol = :value,
    output::Symbol = :value,
    parameter::Symbol = :region_national_supply,
    minimal::Bool = false
)

Calculate the regional national supply for each commodity in each region. The regional national supply is defined as total supply minus net exports minus regional local supply.

WiNDCHousehold.representative_agent_data — Method

representative_agent_data(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts representative agent-related parameters from the regional data table.

    vcat(
        leisure_consumption_elasticity(data),
        average_labor_tax_rate(data),
        labor_supply(data),
        leisure_demand(data),
        table(data, 
            :Personal_Consumption,
            :Household_Interest,
            :Transfer_Payment,
            :Savings,
            :Labor_Endowment;
            normalize=:Use
        )
    )

WiNDCHousehold.saverate_data — Method

saverate_data(HH::HouseholdTable; output = :DataFrame)

Extracts saving rate-related parameters from the household data table.

    table(HH, :Investment_Final_Demand; normalize=:Use)

WiNDCHousehold.sectoral_output — Method

sectoral_output(data::T, output = :DataFrame) where T<:AbstractHouseholdTable

Extracts sectoral output-related parameters from the regional data table.

vcat(
    table(data, 
        :Intermediate_Supply, 
        :Intermediate_Demand, 
        :Labor_Demand, 
        :Capital_Demand;
        normalize = :Use
        ),
    output_tax_rate(data),
    capital_tax_rate(data)
    )

WiNDCHousehold.ssk_data — Method

ssk_data(HH::HouseholdTable; output = :DataFrame)

Extracts steady-state capital stock-related parameters from the household data table.

    table(HH, :Investment_Final_Demand; normalize=:Use)

WiNDCHousehold.tax_rate — Method

tax_rate(
    HH::HouseholdTable; 
    column::Symbol = :value, 
    output::Symbol = :value,
    parameter = :tax_rate,
    minimal::Bool = false
)

Calculate the tax rate for each commodity in each region. The tax rate is defined as the ratio of Tax to Absorption.

Required Arguments

HH::HouseholdTable: The regional data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

parameter::Symbol =:tax_rate`: The name of the parameter column.

minimal::Bool = false: Whether to return a minimal output. If true, only the essential columns are returned: [:col, :region, :year, :parameter, output].

WiNDCHousehold.total_supply — Method

total_supply(data::HouseholdTable; column::Symbol = :value, output::Symbol = :value)

Calculate the total supply for each commodity. The total supply is defined as the sum of:

Intermediate_Supply

Household_Supply

WiNDCHousehold.trans_data — Method

trans_data(HH::HouseholdTable; output = :DataFrame)

Extracts transfer-related parameters from the household data table.

    table(HH, :Government_Final_Demand; normalize=:Use)

WiNDCHousehold.transfer_weights — Method

transfer_weights(raw_data::RawHouseholdData)

Compute transfer weights based on the CPS vs NIPA comparison in raw_data.

Transfer weight is defined as the ratio of NIPA transfer income to CPS transfer income for each transfer category. If the CPS income for a category is zero, the weight from either the Meyer or Rothbaum study is used as a fallback.

Transfer Categories

Category	Source

government benefits: unemployment insurance	hucval

government benefits: social security	hssval

government benefits: social security	hssival

government benefits: social security	hdisval

government benefits: veterans' benefits	hvetval

Meyer and Rothbaum Weights

Source	Literature	Value

hucval	meyer	1/0.679

hssval	meyer	1/0.899

hssival	meyer	1/0.759

hdisval	meyer	1/0.819

hvetval	rothbaum	1/0.679

hwcval	meyer	1 / 0.527

hpawval	meyer	1 / 0.487

hsurval	meyer	1 / 0.908

hedval	rothbaum	1 / 0.804

hcspval	rothbaum	1 / 0.804

hfinval	meyer	1 / 0.539

The values are hard coded based on literature and may need to be updated as new studies emerge. Also, the year is fixed at 2024 for these weights. Need to update for more years in the future.

WiNDCHousehold.windc_vs_nipa_income_categories — Method

windc_vs_nipa_income_categories(
    state_table::WiNDCRegional.State,
    nipa::DataFrame
)

Compare WiNDC labor and capital income to NIPA data.